requirements.txt
Flask==3.0.0
torch
torchvision
transformers
pillow
pytesseract
numpy
gunicorn
python-multipart
requests